Update README.md
This repository contains the codebase for a hybrid deep learning model for earthquake-induced building damage detection using UAV imagery. The proposed STCHMDA-CVT model integrates a custom CNN backbone with a Vision Transformer (ViT) enhanced by a novel Sparse Cross-Attention Hybrid Multi-Dimensional Attention (SCAHMDA) module. It is developed in the context of the 2023 Turkey earthquakes and is trained on the UAVs-TEBDE dataset, which includes annotated aerial images across three damage categories: Cracked, Partially Damaged, and Collapsed. The dataset is available in the Mendeley